+++ Rules/IPv4/Outbound/Software/Nvidia.ps1
@@ -161,7 +161,7 @@ if ([System.Environment]::Is64BitOperatingSystem)
 			Write-Warning -Message "Executable '$NVDisplayExe' was not found, rules for '$NVDisplayExe' won't have any effect"
 
 			Write-Information -Tags "User" -MessageData "INFO: Searched path was: %SystemRoot%\System32\DriverStore\FileRepository"
-			Write-Information -Tags "User" -MessageData "INFO: To fix the problem find '$NVDisplayExe' and adjust the path in $Script and re-run the script"
+			Write-Information -Tags "User" -MessageData "INFO: To fix this problem find '$NVDisplayExe' and adjust the path in $Script and re-run the script"
 		}
 		else
 		{